move min_write_gap to transport layer
